==> 1_data_cleaning.ipynb <==
1. Install and import necessary packages

# Install a-world-of-countries if you don't have it yet (uncomment the next line of code)
#!pip install a-world-of-countries

import pandas as pd
import numpy as np
import awoc #import the a-world-of-countries (awoc) library

2. Load and check dataset

# Load the datasets as a dataframes
df_manifesto = pd.read_csv('RAW_DATA/MPDataset_MPDS2023a.csv')
df_gini = pd.read_csv('RAW_DATA/economic-inequality-gini-index.csv')
df_welfare = pd.read_csv('RAW_DATA/social_welfare_spending.csv')
df_dem_level = pd.read_csv('RAW_DATA/V-Dem-CY-Core-v14.csv')
df_gni = pd.read_csv('RAW_DATA/GNI_34b2ae76-01f3-4a11-82d6-b890757be071_Series - Metadata.csv', skipfooter= 7)

3. Change datatypes where necessary

# Change datatypes in df_manifesto for columns countryname, partyname, partyabbrev, candidatename, corpusversion, datasetversion, and id_perm from object into string.
columns_to_convert = ['countryname', 'partyname', 'partyabbrev', 'candidatename', 'corpusversion', 'datasetversion', 'id_perm']
df_manifesto[columns_to_convert] = df_manifesto[columns_to_convert].astype('string')

# Change datatype in df_manifesto for edate from object into datetime
df_manifesto['edate'] = pd.to_datetime(df_manifesto['edate'], format='%d/%m/%Y')

# Change datatypes in df_gini for columns Entity and Code from object to string
df_gini[['Entity', 'Code']] = df_gini[['Entity', 'Code']].astype('string')

4. Drop possible duplicate rows

df_manifesto = df_manifesto.drop_duplicates()
df_gini = df_gini.drop_duplicates()
df_welfare = df_welfare.drop_duplicates()
df_dem_level = df_dem_level.drop_duplicates()

5. Combine manifesto and gini dataset into one

5.1 Drop unnecessary columns from df_gini

df_gini = df_gini.drop(columns=['Code'])

5.2 Relabel Entity and Year column in df_gini to match with df_manifesto

df_gini = df_gini.rename(columns={"Entity" : "countryname", "Year" : "year"})

5.3 Extract year from edate in df_manifesto

df_manifesto['year'] = df_manifesto['edate'].dt.year

5.4 Merge dataframes based on 'countryname' and 'year' columns

df = pd.merge(df_manifesto, df_gini, on=['countryname', 'year'], how='left')

6. Add welfare state to dataframe

6.1 Reduce welfare dataframe to required columns only (country, year, welfare spending)

df_welfare = df_welfare[['Reference area', 'TIME_PERIOD', 'OBS_VALUE']]

6.2 Relabel columns in df_welfare to match with df

df_welfare.rename(columns={'Reference area': 'countryname', 'TIME_PERIOD': 'year', 'OBS_VALUE': 'welfare spending'}, inplace=True)

6.3 Merge the datasets based on 'countryname' and 'year' columns

df = pd.merge(df, df_welfare, on=['countryname', 'year'], how='left')

7. Add democracy level to dataframe

7.1 Reduce democracy level dataframe to required columns only (country, year, electoral democracy index)

df_dem_level = df_dem_level[['countryname', 'year', 'electoral_dem_ind']]

7.2 Merge the datasets based on 'countryname' and 'year' columns

df = pd.merge(df, df_dem_level, on=['countryname', 'year'], how='left')

8. Add GNI per capita to dataframe

8.1 Reduce GNI dataframe to required columns only (countryname, year, gni)

df_gni = df_gni[['Country Name', 'Time', 'GNI (current US$) [NY.GNP.MKTP.CD]']]

8.2 Relabel columns in df_gni to match with df

df_gni.rename(columns={'Country Name': 'countryname', 'Time': 'year', 'GNI (current US$) [NY.GNP.MKTP.CD]': 'GNI'}, inplace=True)

8.3 Merge the datasets based on 'countryname' and 'year' columns

df = pd.merge(df, df_gni, on=['countryname', 'year'], how='left')

9. Create a list of European countries and create a dataframe with only european countries

# Initialize AWOC class
my_world = awoc.AWOC()

# Create list of European countries and print it
europe_countries = my_world.get_countries_list_of('Europe')
print(europe_countries)

europe_df = df[df['countryname'].isin(europe_countries)]

display(europe_df)

europe_df_2005_2020 = europe_df[(europe_df['edate'].dt.year >= 2005) & (europe_df['edate'].dt.year <= 2020)]
display(europe_df_2005_2020)

# Save dataframe as csv file
europe_df_2005_2020.to_csv('CLEAN_DATA/Data_europe_2005_2020.csv', index=False)

df_dem_level

# Filter rows where the year is 2015 - 2020, saving in new dataframe for chunk selection for the survey
df_2005_2020 = df[(df['edate'].dt.year >= 2005) & (df['edate'].dt.year <= 2020)]

# Filter European countries
europe_2005_2020_df = df_2005_2020[df_2005_2020['countryname'].isin(europe_countries)]
display(europe_2005_2020_df)

# Manifesto selection for survey
survey = europe_2005_2020_df.sample(n=10, random_state=8)
display(survey)

# Filter democracy data and replace data file, file size too large for git otherwise
df_dem_level = df_dem_level[(df_dem_level['year'] >= 2005) & (df_dem_level['year'] <= 2020)]
df_dem_level = df_dem_level[df_dem_level['countryname'].isin(europe_countries)]
df_dem_level.to_csv('RAW_DATA/V-Dem-CY-Core-v14.csv')
