==> 4_interactive_variable.ipynb <==
1. Install and import necessary packages

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns

2. Load and double check dataset

# Load the dataset 
df = pd.read_csv('CLEAN_DATA/2005-2020_percountry_peryear.csv')

display(df)

3. Create interactive variable, representing the interaction between the country features

df['variable_interaction'] = df['rile'] * df['stance_positivity_score'] * df['welfare spending'] * df['electoral_dem_ind'] * df['GNI']

display(df['variable_interaction'])

#save as csv (ps = positivity score)
df.to_csv("CLEAN_DATA/2005-2020_percountry_peryear_final.csv")
