==> 8_survey_chunk_selection.ipynb <==
import random

# define function to collect chunks

def get_chunks(text, maxlength, num_chunks, seed=None):
    random.seed(seed)
    chunks = []
    words = text.split()
    while len(chunks) < num_chunks and len(words) >= maxlength:
        start_index = random.randint(0, len(words) - maxlength)
        end_index = start_index + maxlength
        chunk = ' '.join(words[start_index:end_index])
        chunks.append(chunk)
        del words[start_index:end_index]
    return chunks

file_names = ['14810.txt', '35311.txt', '43901.txt', '51903.txt', '88320.txt', '91440.txt', '92022.txt', '96630.txt']

file_chunks = {}

for file_name in file_names:
    with open('RandomManifestos/Translations/' + file_name, 'r') as file: 
        file_content= file.read()
        chunks = get_chunks(file_content, 500, 5, seed = 42)
        file_chunks[file_name[:-4]] = chunks

output_file = "RandomManifestos/chunks.txt"
with open(output_file, 'w') as file:
    for var_name, chunks in file_chunks.items():
        file.write(f"Variable name: {var_name}\n")
        for i, chunk in enumerate(chunks):
            file.write(f"Chunk {i + 1}:\n{chunk}\n")
        file.write('\n')

print("File chunks have been saved to chunks.txt")
